add missing feilds
